implement IndexSinceLastRunService
