+++ config/default/IndexService.conf.php
@@ -4,8 +4,4 @@
  */
 use oat\tao\model\search\index\IndexService;
 
-return new IndexService([
-    IndexService::OPTION_LASTRUN_STORE => 'YOUR_KEYVALUE_PERSISTENCE',
-    IndexService::OPTION_INDEX_SINCE_LAST_RUN => false,
-
-]);
+return new IndexService();
